requirements.txt
pytelegrambotapi
pydantic
python-dotenv
matplotlib
pandas
async
aiogram
numpy
re